Move those components to globaltypes
