+++ src/applications/dispute-debt/containers/App.tsx
@@ -4,7 +4,7 @@ import { isLoggedIn, selectProfile } from 'platform/user/selectors';
 import { useFeatureToggle } from 'platform/utilities/feature-toggles';
 import { I18nextProvider } from 'react-i18next';
 
-import { RoutedSavableApp, DowntimeNotification } from '../types/platform-components-typed';
+import { RoutedSavableApp, DowntimeNotification } from '../../../../config/globalTypes/platform-components-typed';
 import i18nDebtApp from '../i18n';
 import formConfig from '../config/form';
 import { fetchDebts } from '../actions';
@@ -53,7 +53,6 @@ export const App = ({ children, location }: DisputeDebtProps) => {
 
   return (
     <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
-      {/* @ts-expect-error - HOC-injected props provided automatically by connect() */}
       <DowntimeNotification
         appTitle="dispute debt system"
         dependencies={formConfig.downtime.dependencies}
